Switch to nfs-client-provisioner fork
Switching to my fork of the nfs-client-provisioner.  This creates nfs
directories that do not include the randomized pv name and just the
namespace and pvc name.  It also doesn't delete the NFS directory when
the PVC is deleted.  This is dumb and wouldn't make sense at all unless
you have a pretty static set of apps and want to be able to blow away
the cluster and still have your data in PVs around.